requirements.txt
streamlit
openai
pypdf2
google-generativeai